backend/requirements.txt
asgiref==3.8.1
celery==5.4.0
django==5.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
gunicorn==23.0.0
hiredis==3.1.0
openai==1.63.2
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
rapidfuzz==3.12.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3